+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- added: Notification status analytics param
 - fixed: Some text inputs not selectable for some Samsung devices
 
 ## 4.1.0